Practice Linq 1
The retaurant Host (who is in charge of the waiters, seats people and
takes payments) needs the following information:
Show me the reservations slated for a given Year, Month and Day
(ignoring the cancelled ones). (For sample dates, choose May 29, 2014
and September 20, 1014). - Advanced: Group the reservations by hour of
the day